Play Fair Cipher
Implement the Playfair cipher (in Python). The program should take as input a key
and a file, containing your plaintext/ciphertext (in English only) and operate in two
modes: (a) encryption, and (b) decryption. The output should be a file that contains
the result of encryption/decryption. The key should be entered as a word and your
program should be able to form 5x5 matrix by padding it with remaining letters from
English alphabets. For example: if the key is a word ”crypto”, then the remaining 19
(i.e., 25-6) characters will be ”abdefghiklmnqsuvwxz”. So the full key will be (25 chars):
cryptoabdfghiklmnopquvwxz, which you have to internally arrange as 5x5 matrix.
Note: Don’t use any library that implements Playfair. Provide a README file that
explain clearly how to run your code. In case you are wondering if you can use other
language to implement, Python code is preferred over, say, C for this assignment.